- (config*, agent/mibgroup/*.h):   - new config macro:  config_arch_require(arch, module).   - moved arch dependant modules (memory, hpux, kernel_sunos5) loads     from the configure into the appropriate .h files. - (interfaces.c, var_route.c):   - typecast klookup calls for freebsd2.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@979 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/var_route.c
@@ -569,7 +569,7 @@ struct radix_node *pt;
 #endif
       name[sizeof (name)-1] = '\0';
 #else
-      klookup( ifnet.if_name, name, sizeof name);
+      klookup((unsigned long)ifnet.if_name, name, sizeof name);
       name[sizeof (name) - 1] = '\0';
       cp = (char *) index(name, '\0');
       string_append_int (cp, ifnet.if_unit);
